+++ src/models/scripts/contributors/contributor.ts
@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 export interface Contributor {
     login: string;
     id: number;